requirements.txt
asgiref==3.8.1
Django==5.2.3
sqlparse==0.5.3
psycopg2-binary==2.9.10
djangorestframework>=3.15.0
django-simple-history>=3.4.0
django-filter>=24.1
drf-yasg>=1.21.7
